requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
langchain-google-genai
chromadb
python-dotenv